+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 
 * Remove code needing c++17, and add the c++11 compile flag (#516)
 * Don't skip frames in Vulkan for now (#492)
+* Immediately detach all created threads, and handle joinable state manually
 
 ### Fixed